HOME OFFICE SECURITY SYSTEM using PIC Microcontroller 16F877A
DESIGN
The main control circuit is around the microcontroller PIC 16F877A. Home
/office security system consists of a sensor circuit to identify the
interruption in restricted or secure place. If an interruption is
detected a HIGH pulse is sent to a port of the microcontroller. That is
identified as an insecure condition. So the controller then sends
message of insecure condition and also calls to a predetermined phone
number. The optional provision is included to send a message in a
specified format to command for unlocking the locked gate.